Move two-part-tariff files to billing module folder
