+++ test/modules/billing/jobs/two-part-tariff/prepare-charge-elements.js
@@ -5,7 +5,7 @@ const {
   getTptChargeElements,
   prepareChargeElementData,
   sortChargeElementsForMatching
-} = require('../../../../src/modules/charging/lib/prepare-charge-elements');
+} = require('../../../../../src/modules/billing/jobs/two-part-tariff/prepare-charge-elements');
 
 experiment('modules/charging/lib/prepare-charge-elements', async () => {
   experiment('.getTptChargeElements', async () => {
